move some things around and fix imports
