+++ modules/app/components/layout/header/AccountBox.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 import { Flex, Text, Box, Button, Link as ExternalLink } from 'theme-ui';
 import { Icon } from '@makerdao/dai-ui-icons';
-import { ConnectorName } from 'lib/web3react';
 import { useBreakpointIndex } from '@theme-ui/match-media';
 
 import { formatAddress, getEtherscanLink } from 'lib/utils';
 import { getNetwork } from 'lib/maker';
 import AddressIcon from 'modules/address/components/AddressIcon';
+import { ConnectorName } from 'modules/web3/types/connectors';
 
 type Props = {
   address: string;
