+++ modules/app/components/layout/header/AccountSelect.tsx
@@ -19,12 +19,13 @@ import NetworkAlertModal from './NetworkAlertModal';
 import useAccountsStore from 'modules/app/stores/accounts';
 import { AbstractConnector } from '@web3-react/abstract-connector';
 import ConnectWalletButton from 'modules/web3/components/ConnectWalletButton';
-import { getLibrary } from 'modules/web3/helpers';
+import { chainIdToNetworkName, getLibrary } from 'modules/web3/helpers';
 import { useEagerConnect } from 'modules/web3/hooks/useEagerConnect';
-import { ConnectorName, SUPPORTED_WALLETS, chainIdToNetworkName } from 'modules/web3/web3.constants';
+import { SUPPORTED_WALLETS } from 'modules/web3/wallets';
 import { useContext } from 'react';
 import { AnalyticsContext } from 'modules/app/client/analytics/AnalyticsContext';
 import Tooltip from 'modules/app/components/Tooltip';
+import { ConnectorName } from 'modules/web3/types/connectors';
 
 export type ChainIdError = null | 'network mismatch' | 'unsupported network';
