+++ modules/web3/types/connectors.d.ts
@@ -1,0 +1,2 @@
+export type ConnectorName = 'MetaMask' | 'WalletConnect' | 'Coinbase Wallet';
+//| 'Trezor' | 'Ledger';
