+++ modules/web3/types/wallets.d.ts
@@ -1,0 +1,7 @@
+import { AbstractConnector } from '@web3-react/abstract-connector';
+import { ConnectorName } from './connectors';
+
+export interface WalletInfo {
+  connector: AbstractConnector;
+  name: ConnectorName;
+}
