+++ modules/web3/web3.constants.ts
@@ -1,11 +1,4 @@
-import { AbstractConnector } from '@web3-react/abstract-connector';
-import { InjectedConnector } from '@web3-react/injected-connector';
-import { WalletConnectConnector, WalletConnectConnectorArguments } from '@web3-react/walletconnect-connector';
-import { WalletLinkConnector } from '@web3-react/walletlink-connector';
 import { config } from 'lib/config';
-// import { ConnectorName } from 'lib/web3react';
-
-const POLLING_INTERVAL = 12000;
 
 export enum SupportedConnectors {
   METAMASK = 'metamask',
@@ -81,61 +74,3 @@ export const NETWORK_URLS = {
 };
 export const DEFAULT_NETWORK = SupportedNetworks.MAINNET;
 export const DEFAULT_NODE_PROVIDER = NodeProviders.INFURA;
-
-export const networkToRpc = (
-  network: string = DEFAULT_NETWORK,
-  nodeProvider: string = DEFAULT_NODE_PROVIDER
-): string => NETWORK_URLS[network][nodeProvider];
-
-export const chainIdToNetworkName = (chainId: number): SupportedNetworks => {
-  for (const chain in CHAIN_INFO) {
-    if (CHAIN_INFO[chain].chainId === chainId) {
-      return chain as SupportedNetworks;
-    }
-  }
-  throw new Error(`Unsupported chain id ${chainId}`);
-};
-export const injectedConnector = new InjectedConnector({
-  supportedChainIds: [
-    CHAIN_INFO[SupportedNetworks.MAINNET].chainId,
-    CHAIN_INFO[SupportedNetworks.GOERLI].chainId,
-    CHAIN_INFO[SupportedNetworks.KOVAN].chainId,
-    CHAIN_INFO[SupportedNetworks.TESTNET].chainId
-  ]
-});
-
-export const walletConnectConnector = new WalletConnectConnector({
-  rpc: { 1: networkToRpc(SupportedNetworks.MAINNET, NodeProviders.INFURA) },
-  bridge: 'https://bridge.walletconnect.org',
-  qrcode: true,
-  pollingInterval: POLLING_INTERVAL
-} as WalletConnectConnectorArguments);
-
-export const walletLinkConnector = new WalletLinkConnector({
-  url: networkToRpc(SupportedNetworks.MAINNET, 'infura'),
-  appName: 'vote.makerdao.com'
-});
-
-/**Wallets */
-export type ConnectorName = 'MetaMask' | 'WalletConnect' | 'Coinbase Wallet';
-//| 'Trezor' | 'Ledger';
-
-interface WalletInfo {
-  connector: AbstractConnector;
-  name: ConnectorName;
-}
-
-export const SUPPORTED_WALLETS: { [connector in SupportedConnectors]: WalletInfo } = {
-  [SupportedConnectors.METAMASK]: {
-    connector: injectedConnector,
-    name: 'MetaMask'
-  },
-  [SupportedConnectors.WALLET_CONNECT]: {
-    connector: walletConnectConnector,
-    name: 'WalletConnect'
-  },
-  [SupportedConnectors.COINBASE_WALLET]: {
-    connector: walletLinkConnector,
-    name: 'Coinbase Wallet'
-  }
-};
