+++ pages/api/address/[address]/delegated-to.ts
@@ -1,7 +1,7 @@
 import invariant from 'tiny-invariant';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { isSupportedNetwork } from 'lib/maker/index';
-import { DEFAULT_NETWORK } from 'lib/constants';
+import { DEFAULT_NETWORK } from 'modules/web3/web3.constants';
 import withApiHandler from 'lib/api/withApiHandler';
 import { resolveENS } from 'modules/web3/ens';
 import { fetchDelegatedTo } from 'modules/delegates/api/fetchDelegatedTo';
