+++ pages/api/comments/executive/add/[address].ts
@@ -5,8 +5,9 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { connectToDatabase } from 'lib/api/utils';
 import withApiHandler from 'lib/api/withApiHandler';
 import { config } from 'lib/config';
-import { DEFAULT_NETWORK, SupportedNetworks } from 'lib/constants';
+import { SupportedNetworks } from 'lib/constants';
 import { ExecutiveComment, ExecutiveCommentsRequestBody } from 'modules/comments/types/executiveComment';
+import { DEFAULT_NETWORK } from 'modules/web3/web3.constants';
 
 export default withApiHandler(
   async (req: NextApiRequest, res: NextApiResponse) => {
