+++ pages/api/executive/supporters.ts
@@ -2,7 +2,7 @@ import invariant from 'tiny-invariant';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import getMaker, { isSupportedNetwork } from 'lib/maker';
-import { DEFAULT_NETWORK } from 'lib/constants';
+import { DEFAULT_NETWORK } from 'modules/web3/web3.constants';
 import withApiHandler from 'lib/api/withApiHandler';
 import { fetchDelegates } from 'modules/delegates/api/fetchDelegates';
