+++ pages/api/polling/tally/[poll-id].ts
@@ -1,8 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import withApiHandler from 'lib/api/withApiHandler';
-import { DEFAULT_NETWORK, SupportedNetworks } from 'lib/constants';
+import { SupportedNetworks } from 'lib/constants';
 import { getPollTally } from 'modules/polling/helpers/getPollTaly';
 import { getPollById } from 'modules/polling/api/fetchPolls';
+import { DEFAULT_NETWORK } from 'modules/web3/web3.constants';
 
 // Returns a PollTally given a pollID
